Moved mutations and queries to own folders
